+++ igi1conv/cmd_lightmap.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int cmd_lightmap(int argc, char** argv);